+++ CHANGELOG.md
@@ -115,6 +115,7 @@
 - Add streaming support for the video content type @robgietema
 - Add exportimport script to migrate data @robgietema
 - Add context navigation endpoint @robgietema
+- Add depth option to navigation endpoint @robgietema
 
 ### Bugfix